Major API Updates.
The codebase is now base-of-svn-checkout aware. It uses mostly relative paths, defined in the new Paths class.
Much of launchHack.py's code was factored out into a TargetLauncher class.

The API has been divided internally. There is now an 'inside' and 'outside' API. The outside API is used for things like the launcher. The inside API is used for things liek detours and memory manipulation once inside the process. The division here is one of conceptual process boundry - once the DLL loads your file, you should be using the inside API, as you are now inside the target.

pyDetour no longer calls your python file directly. It first calls bootstrap.py to setup the initial enviornment.

launchHack.py, when run directly, pretends it was imported instead. This makes it looks like part of the pyHack package to Python's import mechanism, which allows package based imports to work well in the outside enviornment. It also stores the names of imported classed correctly now (ex: pyhack.detour_api.targetDef vs detour_api.targetDef) which means things can be pickled easier and we don't need to pollute sys.path as much.

There is much more to come, such as more extensive documentation, and a better, more sane PatchManagement system. As it stands now, the eventual destination python code is still effectviely main(), which is a problem, because we want to be able to call multiple 'patch plugins' in succession. The proper way to do this is to pass through the enviornment a PatchList of some sort instead of the single target file, and then loop through them and imp.load_source them in bootstrap.py.